rename directory to `bootstrap`
